Index date_created field as text and use for display
DLTP-257

There were some problems with indexing the date_created field.
The indexing worked fine for Article and Image
but not for Dataset and Document.
The only difference I noticed was the Metadata datastreams for Document
and Datasets was extending Generic Works Rdf datastream.
Where as the Article and Image didn't.
But when I moved the date_created from dataset and document to
GenericWork, it started indexing fine